pyproject.toml
[project]
name = "botfoochan"
version = "0.3.0"
authors = [{ name = "StarHeartHunt", email = "starheart233@gmail.com" }]
description = "Multi channel Mooncell bot"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "diro-py ~=0.1",
  "matplotlib ~=3.8",
  "playwright ~=1.38",
  "nonebot2[httpx,fastapi,websockets] ~=2.1",
  "nonebot-adapter-qq ~=1.1",
  "nonebot-adapter-feishu ~=2.3",
  "nonebot-adapter-onebot ~=2.3",
  "nonebot-plugin-sentry ~=2.0",
  "nonebot-plugin-alconna ~=0.38",
  "nonebot-plugin-filehost ~=1.1",
  "nonebot-plugin-shorturl ~=0.4",
  "nonebot-plugin-apscheduler ~=0.4",
  "nonebot-bison >=0.9.6rc2, <1.0",
  "nonebot-plugin-treehelp ~=0.3",
  "openai ~=1.61",
]

[dependency-groups]
dev = ["ruff ~= 0.9", "pre-commit ~= 4.1"]
deploy = ["gunicorn ~= 23.0", "uvicorn ~= 0.18"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
  "F",     # pyflakes
  "W",     # pycodestyle warnings
  "E",     # pycodestyle errors
  "I",     # isort
  "UP",    # pyupgrade
  "ASYNC", # flake8-async
  "C4",    # flake8-comprehensions
  "DTZ",   # flake8-datetimez
  "T10",   # flake8-debugger
  "T20",   # flake8-print
  "PYI",   # flake8-pyi
  "PT",    # flake8-pytest-style
  "Q",     # flake8-quotes
  "TC",    # flake8-type-checking
  "TID",   # flake8-tidy-imports
  "RUF",   # Ruff-specific rules
]
ignore = [
  "E402",   # module-import-not-at-top-of-file
  "UP037",  # quoted-annotation
  "RUF001", # ambiguous-unicode-character-string
  "RUF002", # ambiguous-unicode-character-docstring
  "RUF003", # ambiguous-unicode-character-comment
]

[tool.ruff.lint.isort]
force-sort-within-sections = true

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
pythonPlatform = "All"

typeCheckingMode = "standard"
reportShadowedImports = false
disableBytesTypePromotions = true
